refactor(overlay): split service.rs into per-backend modules
Pure code motion: x11.rs, wayland.rs, gnome.rs and shared render.rs,
matching the one-file-per-backend layout of src/window/. spawn_overlay
stays the sole public export.
