+++ src/overlay/gnome.rs
@@ -1,0 +1,112 @@
+//! GNOME overlay D-Bus broadcaster.
+
+use std::time::Duration;
+
+use tokio::sync::watch;
+use tracing::info;
+
+use crate::State;
+
+use super::service::OverlayError;
+
+pub(super) fn is_gnome_desktop() -> bool {
+    std::env::var("XDG_CURRENT_DESKTOP")
+        .map(|value| {
+            value
+                .split(':')
+                .any(|part| part.eq_ignore_ascii_case("GNOME"))
+        })
+        .unwrap_or(false)
+}
+
+pub(super) async fn run_gnome_broadcaster(
+    mut state_rx: watch::Receiver<State>,
+    level_rx: watch::Receiver<f32>,
+    theme: String,
+) -> Result<(), OverlayError> {
+    // Custom themes don't sync over D-Bus for v1 — the GNOME extension only
+    // knows the named themes it ships. Fall back to "ember" so the bar
+    // colors remain sensible.
+    let advertised_theme = match theme.as_str() {
+        "carbon" | "cyan" | "ember" => theme.clone(),
+        _ => "ember".to_string(),
+    };
+
+    let conn = zbus::connection::Builder::session()?
+        .serve_at("/org/whisrs/Overlay", GnomeOverlayBus)?
+        .name("org.whisrs.Overlay")?
+        .build()
+        .await?;
+
+    info!("GNOME overlay D-Bus broadcaster started");
+    emit_gnome_theme(&conn, &advertised_theme).await?;
+    let initial_state = *state_rx.borrow();
+    emit_gnome_state(&conn, initial_state).await?;
+    let initial_level = *level_rx.borrow();
+    emit_gnome_level(&conn, initial_level).await?;
+
+    // Emit level at a steady ~30 Hz so the GNOME extension always has fresh
+    // data, regardless of how the watch channel coalesces updates.
+    let mut level_interval = tokio::time::interval(Duration::from_millis(33));
+    level_interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
+
+    loop {
+        tokio::select! {
+            changed = state_rx.changed() => {
+                if changed.is_err() {
+                    break;
+                }
+                let state = *state_rx.borrow();
+                emit_gnome_state(&conn, state).await?;
+            }
+            _ = level_interval.tick() => {
+                let level = *level_rx.borrow();
+                emit_gnome_level(&conn, level).await?;
+            }
+        }
+    }
+
+    Ok(())
+}
+
+async fn emit_gnome_state(conn: &zbus::Connection, state: State) -> zbus::Result<()> {
+    conn.emit_signal(
+        None::<&str>,
+        "/org/whisrs/Overlay",
+        "org.whisrs.Overlay",
+        "StateChanged",
+        &(state.to_string()),
+    )
+    .await
+}
+
+async fn emit_gnome_level(conn: &zbus::Connection, level: f32) -> zbus::Result<()> {
+    conn.emit_signal(
+        None::<&str>,
+        "/org/whisrs/Overlay",
+        "org.whisrs.Overlay",
+        "LevelChanged",
+        &level.clamp(0.0, 1.0),
+    )
+    .await
+}
+
+async fn emit_gnome_theme(conn: &zbus::Connection, theme: &str) -> zbus::Result<()> {
+    conn.emit_signal(
+        None::<&str>,
+        "/org/whisrs/Overlay",
+        "org.whisrs.Overlay",
+        "ThemeChanged",
+        &theme,
+    )
+    .await
+}
+
+struct GnomeOverlayBus;
+
+#[zbus::interface(name = "org.whisrs.Overlay")]
+impl GnomeOverlayBus {
+    fn ping(&self) -> &'static str {
+        "ok"
+    }
+}
