+++ src/overlay/mod.rs
@@ -3,7 +3,15 @@
 //! The overlay is optional and has native Wayland layer-shell and X11 backends.
 
 #[cfg(feature = "overlay")]
+mod gnome;
+#[cfg(feature = "overlay")]
+mod render;
+#[cfg(feature = "overlay")]
 mod service;
+#[cfg(feature = "overlay")]
+mod wayland;
+#[cfg(feature = "overlay")]
+mod x11;
 
 #[cfg(feature = "overlay")]
 pub use service::spawn_overlay;
