+++ src/overlay/wayland.rs
@@ -1,0 +1,286 @@
+//! Wayland layer-shell overlay backend.
+
+use std::sync::mpsc;
+use std::time::Duration;
+
+use smithay_client_toolkit::{
+    compositor::{CompositorHandler, CompositorState},
+    delegate_compositor, delegate_layer, delegate_output, delegate_registry, delegate_shm,
+    output::{OutputHandler, OutputState},
+    registry::{ProvidesRegistryState, RegistryState},
+    registry_handlers,
+    shell::{
+        wlr_layer::{
+            Anchor, KeyboardInteractivity, Layer, LayerShell, LayerShellHandler, LayerSurface,
+            LayerSurfaceConfigure,
+        },
+        WaylandSurface,
+    },
+    shm::{slot::SlotPool, Shm, ShmHandler},
+};
+use tiny_skia::{Pixmap, PremultipliedColorU8};
+use tracing::{info, warn};
+use wayland_client::{
+    globals::registry_queue_init,
+    protocol::{wl_output, wl_region, wl_shm, wl_surface},
+    Connection as WaylandConnection, Dispatch, QueueHandle,
+};
+
+use crate::{OverlayConfig, State};
+
+use super::render::{OverlayRenderer, Theme, BOTTOM_MARGIN, FRAME_MS};
+use super::service::OverlayError;
+
+pub(super) fn run_overlay(
+    state_rx: mpsc::Receiver<State>,
+    level_rx: mpsc::Receiver<f32>,
+    config: OverlayConfig,
+) -> Result<(), OverlayError> {
+    let width = config.clamped_width();
+    let height = config.clamped_height();
+    let theme = Theme::from_config(&config);
+
+    let conn = WaylandConnection::connect_to_env()?;
+    let (globals, mut event_queue) = registry_queue_init(&conn)?;
+    let qh = event_queue.handle();
+
+    let compositor = CompositorState::bind(&globals, &qh)?;
+    let layer_shell = LayerShell::bind(&globals, &qh)?;
+    let shm = Shm::bind(&globals, &qh)?;
+
+    let surface = compositor.create_surface(&qh);
+    let layer =
+        layer_shell.create_layer_surface(&qh, surface, Layer::Overlay, Some("whisrs"), None);
+    layer.set_anchor(Anchor::BOTTOM);
+    layer.set_margin(0, 0, BOTTOM_MARGIN, 0);
+    layer.set_exclusive_zone(0);
+    layer.set_keyboard_interactivity(KeyboardInteractivity::None);
+    layer.set_size(width, height);
+
+    // Make the transparent overlay non-interactive so it never blocks clicks.
+    let input_region = compositor.wl_compositor().create_region(&qh, ());
+    layer.set_input_region(Some(&input_region));
+    input_region.destroy();
+
+    layer.commit();
+
+    let pool = SlotPool::new((width * height * 4) as usize, &shm)?;
+    let mut overlay = Overlay {
+        registry_state: RegistryState::new(&globals),
+        output_state: OutputState::new(&globals, &qh),
+        shm,
+        pool,
+        layer,
+        renderer: OverlayRenderer::new(state_rx, level_rx, width, height, theme)?,
+        exit: false,
+        first_configure: true,
+    };
+
+    info!("recording overlay started");
+    while !overlay.exit {
+        overlay.renderer.apply_state_updates();
+        if overlay.renderer.disconnected {
+            break;
+        }
+        event_queue.blocking_dispatch(&mut overlay)?;
+    }
+
+    Ok(())
+}
+
+struct Overlay {
+    registry_state: RegistryState,
+    output_state: OutputState,
+    shm: Shm,
+    pool: SlotPool,
+    layer: LayerSurface,
+    renderer: OverlayRenderer,
+    exit: bool,
+    first_configure: bool,
+}
+
+impl Overlay {
+    fn draw(&mut self, qh: &QueueHandle<Self>) {
+        let width = self.renderer.width;
+        let height = self.renderer.height;
+        let stride = width as i32 * 4;
+
+        self.renderer.draw_frame();
+
+        let Ok((buffer, canvas)) = self.pool.create_buffer(
+            width as i32,
+            height as i32,
+            stride,
+            wl_shm::Format::Argb8888,
+        ) else {
+            warn!("failed to allocate overlay buffer");
+            return;
+        };
+        copy_pixmap_to_argb8888(&self.renderer.pixmap, canvas);
+
+        self.layer
+            .wl_surface()
+            .damage_buffer(0, 0, width as i32, height as i32);
+        self.layer
+            .wl_surface()
+            .frame(qh, self.layer.wl_surface().clone());
+        if let Err(e) = buffer.attach_to(self.layer.wl_surface()) {
+            warn!("failed to attach overlay buffer: {e}");
+            return;
+        }
+        self.layer.commit();
+
+        std::thread::sleep(Duration::from_millis(FRAME_MS));
+    }
+}
+
+/// tiny-skia stores premultiplied RGBA bytes; the wl_shm Argb8888 format on
+/// little-endian systems is BGRA in memory. Convert by swapping R and B.
+/// Both formats use premultiplied alpha so no math is needed beyond the swap.
+fn copy_pixmap_to_argb8888(pixmap: &Pixmap, canvas: &mut [u8]) {
+    let src = pixmap.pixels();
+    debug_assert_eq!(src.len() * 4, canvas.len());
+    for (i, px) in src.iter().enumerate() {
+        let dst = &mut canvas[i * 4..i * 4 + 4];
+        let pre: PremultipliedColorU8 = *px;
+        dst[0] = pre.blue();
+        dst[1] = pre.green();
+        dst[2] = pre.red();
+        dst[3] = pre.alpha();
+    }
+}
+
+impl CompositorHandler for Overlay {
+    fn scale_factor_changed(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _surface: &wl_surface::WlSurface,
+        _new_factor: i32,
+    ) {
+    }
+
+    fn transform_changed(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _surface: &wl_surface::WlSurface,
+        _new_transform: wl_output::Transform,
+    ) {
+    }
+
+    fn frame(
+        &mut self,
+        _conn: &WaylandConnection,
+        qh: &QueueHandle<Self>,
+        _surface: &wl_surface::WlSurface,
+        _time: u32,
+    ) {
+        self.draw(qh);
+    }
+
+    fn surface_enter(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _surface: &wl_surface::WlSurface,
+        _output: &wl_output::WlOutput,
+    ) {
+    }
+
+    fn surface_leave(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _surface: &wl_surface::WlSurface,
+        _output: &wl_output::WlOutput,
+    ) {
+    }
+}
+
+impl OutputHandler for Overlay {
+    fn output_state(&mut self) -> &mut OutputState {
+        &mut self.output_state
+    }
+
+    fn new_output(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _output: wl_output::WlOutput,
+    ) {
+    }
+
+    fn update_output(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _output: wl_output::WlOutput,
+    ) {
+    }
+
+    fn output_destroyed(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _output: wl_output::WlOutput,
+    ) {
+    }
+}
+
+impl LayerShellHandler for Overlay {
+    fn closed(
+        &mut self,
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+        _layer: &LayerSurface,
+    ) {
+        self.exit = true;
+    }
+
+    fn configure(
+        &mut self,
+        _conn: &WaylandConnection,
+        qh: &QueueHandle<Self>,
+        _layer: &LayerSurface,
+        _configure: LayerSurfaceConfigure,
+        _serial: u32,
+    ) {
+        if self.first_configure {
+            self.first_configure = false;
+            self.draw(qh);
+        }
+    }
+}
+
+impl ShmHandler for Overlay {
+    fn shm_state(&mut self) -> &mut Shm {
+        &mut self.shm
+    }
+}
+
+impl Dispatch<wl_region::WlRegion, ()> for Overlay {
+    fn event(
+        _state: &mut Self,
+        _proxy: &wl_region::WlRegion,
+        _event: wl_region::Event,
+        _data: &(),
+        _conn: &WaylandConnection,
+        _qh: &QueueHandle<Self>,
+    ) {
+    }
+}
+
+delegate_compositor!(Overlay);
+delegate_output!(Overlay);
+delegate_shm!(Overlay);
+delegate_layer!(Overlay);
+delegate_registry!(Overlay);
+
+impl ProvidesRegistryState for Overlay {
+    fn registry(&mut self) -> &mut RegistryState {
+        &mut self.registry_state
+    }
+
+    registry_handlers![OutputState];
+}
